+++ packages/atlas/src/views/studio/YppDashboard/tabs/YppDashboardReferralsTab/YppDashboardReferralsTab.tsx
@@ -1,41 +1,20 @@
-import { useMemo } from 'react'
-import { useQuery } from 'react-query'
-
-import { axiosInstance } from '@/api/axios'
 import { SvgActionLinkUrl } from '@/assets/icons'
 import { EmptyFallback } from '@/components/EmptyFallback'
 import { GridItem, LayoutGrid } from '@/components/LayoutGrid'
-import { YppReferral, YppReferralTable } from '@/components/YppReferralTable/YppReferralTable'
+import { YppReferralTable } from '@/components/YppReferralTable/YppReferralTable'
 import { ReferralLinkButton } from '@/components/_ypp/ReferralLinkButton'
 import { YppSuspendedBanner } from '@/components/_ypp/YppSuspendedBanner'
 import { atlasConfig } from '@/config'
-import { useUser } from '@/providers/user/user.hooks'
-import { YppSyncedChannel } from '@/views/global/YppLandingView/YppLandingView.types'
+import { useYppReferralPagination } from '@/hooks/useYppReferralPagination'
 
 import { FallbackContainer } from '../YppDashboardTabs.styles'
 
-const YPP_SYNC_URL = atlasConfig.features.ypp.youtubeSyncApiUrl
+const TILES_PER_PAGE = 10
 
 export const YppDashboardReferralsTab = () => {
   const yppSuspended = atlasConfig.features.ypp.suspended
-  const { channelId } = useUser()
-  const { isLoading, data } = useQuery(
-    ['referralsTable', channelId],
-    () => axiosInstance.get<YppSyncedChannel[]>(`${YPP_SYNC_URL}/channels/${channelId}/referrals`),
-    { enabled: !!channelId }
-  )
-
-  const mappedData: YppReferral[] = useMemo(
-    () =>
-      data?.data.map((channelData) => {
-        return {
-          date: new Date(channelData.createdAt),
-          channelId: String(channelData.joystreamChannelId),
-          status: channelData.yppStatus,
-        }
-      }) ?? [],
-    [data?.data]
-  )
+  const { isLoading, yppReferrals, currentPage, setCurrentPage, perPage, setPerPage, totalCount } =
+    useYppReferralPagination({ initialPageSize: TILES_PER_PAGE })
 
   return (
     <LayoutGrid>
@@ -45,7 +24,7 @@ export const YppDashboardReferralsTab = () => {
         </GridItem>
       )}
       <GridItem colSpan={{ base: 12 }}>
-        {!isLoading && !mappedData?.length ? (
+        {!isLoading && totalCount === 0 ? (
           <FallbackContainer>
             <EmptyFallback
               title="No referred users yet"
@@ -55,7 +34,17 @@ export const YppDashboardReferralsTab = () => {
             />
           </FallbackContainer>
         ) : (
-          <YppReferralTable data={mappedData} isLoading={isLoading} />
+          <YppReferralTable
+            data={yppReferrals}
+            isLoading={isLoading}
+            pagination={{
+              page: currentPage,
+              setPerPage,
+              totalCount,
+              itemsPerPage: perPage,
+              onChangePage: setCurrentPage,
+            }}
+          />
         )}
       </GridItem>
     </LayoutGrid>